requirements.txt
streamlit
pandas
numpy
statsmodels
plotly
websocket-client
